+++ src/axom/inlet/ConduitReader.hpp
@@ -76,6 +76,11 @@ public:
   ReaderResult getStringMap(const std::string& id,
                             std::unordered_map<VariantKey, std::string>& values) override;
 
+  ReaderResult getVariantMap(const std::string& id,
+                             std::unordered_map<int, VariantValue>& values) override;
+  ReaderResult getVariantMap(const std::string& id,
+                             std::unordered_map<VariantKey, VariantValue>& values) override;
+
   ReaderResult getIndices(const std::string& id, std::vector<int>& indices) override;
   ReaderResult getIndices(const std::string& id, std::vector<VariantKey>& indices) override;
 
@@ -110,12 +115,14 @@ private:
   ReaderResult getValue(const conduit::Node* node, std::string& value);
   ReaderResult getValue(const conduit::Node* node, double& value);
   ReaderResult getValue(const conduit::Node* node, bool& value);
+  ReaderResult getValue(const conduit::Node* node, VariantValue& value);
 
   template <typename T>
   ReaderResult getDictionary(const std::string& id, std::unordered_map<VariantKey, T>& values);
 
   template <typename T>
   ReaderResult getArray(const std::string& id, std::unordered_map<int, T>& values);
+  ReaderResult getVariantArray(const std::string& id, std::unordered_map<int, VariantValue>& values);
   conduit::Node m_root;
   const std::string m_protocol;
 };